completed decomposition
