+++ client/src/components/LibraryCatalogList/LibraryCatalogList.scss
@@ -1,0 +1,11 @@
+.Library__Catalog-List {
+  // border: 1px solid black;
+  padding: 20px;
+  text-align: center;
+}
+
+.Library__Catalog-List-Zero {
+  text-align: center;
+  margin-top: 30px;
+  font-size: 28px;
+}
