--- client/src/components/LibraryPage/LibraryPage.scss
@@ -1,141 +1,0 @@
-.Library__Page-Container {
-  max-width: 75%;
-  margin: 0 auto;
-  font-family: "Roboto", sans-serif;
-  text-align: center;
-  font-weight: 300;
-  h1 {
-    font-size: 35px;
-    font-weight: 400;
-  }
-  hr {
-    margin-bottom: 30px;
-    margin-top: 30px;
-  }
-
-  h2 {
-    font-size: 23px;
-    margin-bottom: 30px;
-  }
-}
-
-.Library__Page-Descr-Bold {
-  margin-top: 30px;
-  margin-bottom: 30px;
-  font-weight: 300;
-}
-
-.Library__Page-Button-Group {
-  margin-bottom: 30px;
-  button {
-    width: 110px;
-    height: 30px;
-    font-family: "Roboto", sans-serif;
-    font-weight: 300;
-    background-color: #d0d0e5;
-    border: 1px solid #d9d9d9;
-    border-radius: 3px;
-    transition: 0.5s linear;
-    &:hover {
-      background-color: #ababc9;
-    }
-  }
-}
-
-.Library__Page-Button {
-  width: 60%;
-  height: 50px;
-  border: 1px solid white;
-  margin-top: 22px;
-  display: inline-block;
-  cursor: pointer;
-  color: #000000;
-  text-transform: uppercase;
-  font-size: 12px;
-  letter-spacing: 0.15rem;
-  transition: all 0.5s;
-  position: relative;
-  overflow: hidden;
-  z-index: 1;
-  &:after {
-    content: "";
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    background-color: #fff;
-    z-index: -2;
-    //   border: 1px solid white;
-  }
-  &:before {
-    content: "";
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 0%;
-    height: 100%;
-    background-color: darken(white, 15%);
-    transition: all 0.3s;
-    z-index: -1;
-    //   border: 1px solid white;
-  }
-  &:hover {
-    color: #fff;
-    &:before {
-      width: 100%;
-    }
-  }
-}
-
-.Libray__Page-Users {
-  max-width: 75%;
-  margin: 0 auto;
-  h1 {
-    margin-bottom: 30px;
-  }
-  p {
-    color: #000000;
-    margin-top: 10px;
-  }
-}
-
-.Libray__Page-User {
-  border: 1px solid black;
-  padding: 20px;
-  margin: 10px;
-  background-color: #f6f6f6;
-  transition: 0.3s all;
-  &:hover {
-    background-color: #d0d0e5;
-  }
-}
-
-.Libray__Page-User-Button {
-  width: 300px;
-  height: 30px;
-  font-family: "Roboto", sans-serif;
-  font-weight: 300;
-  background-color: #c5c5c5;
-  border: 1px solid #676767;
-  color: #fff;
-  border-radius: 3px;
-  transition: 0.5s linear;
-  margin-top: 10px;
-}
-
-.Library__Page-AddUser {
-  margin-top: 30px;
-}
-
-.Library__Page-Static {
-  display: flex;
-  justify-content: center;
-  gap: 80px;
-
-  h3 {
-    font-size: 21px;
-  }
-}
-
-.Library__Page-Main-Title {
-  margin-bottom: 30px;
-}
