+++ client/src/components/LibraryPageListTwo/LibraryPageListTwo.scss
@@ -1,0 +1,23 @@
+.Library__Page-User {
+  border: 1px solid black;
+  padding: 20px;
+  margin: 10px;
+  background-color: #f6f6f6;
+  transition: 0.3s all;
+  &:hover {
+    background-color: #d0d0e5;
+  }
+}
+
+.Library__Page-User-Button {
+  width: 300px;
+  height: 30px;
+  font-family: "Roboto", sans-serif;
+  font-weight: 300;
+  background-color: #c5c5c5;
+  border: 1px solid #676767;
+  color: #fff;
+  border-radius: 3px;
+  transition: 0.5s linear;
+  margin-top: 10px;
+}
