+++ client/src/components/MyBooks/MyBooksList.scss
@@ -1,0 +1,17 @@
+.MyBooks__SubTitle {
+  text-align: center;
+  margin-top: 30px;
+  font-size: 28px;
+}
+
+.MyBook__List {
+  max-width: 70%;
+  margin: 0 auto;
+  border: 1px solid black;
+  padding: 30px;
+  margin-bottom: 20px;
+}
+
+.MyBook__List-Descr {
+  text-align: center;
+}
