--- client/src/components/Pagination/Pagination.jsx
@@ -1,24 +1,0 @@
-import "./Pagination.scss";
-import { useAction } from "../../store";
-
-const Pagination = ({ currentPage }) => {
-  const pages = [1, 2, 3, 4, 5];
-
-  const { setCurrentPage } = useAction();
-
-  return (
-    <div className="pages">
-      {pages.map((page, index) => (
-        <span
-          className={currentPage == page ? "current__page" : "page"}
-          key={index}
-          onClick={() => setCurrentPage(page)}
-        >
-          {page}
-        </span>
-      ))}
-    </div>
-  );
-};
-
-export default Pagination;
