+++ client/src/components/UI/Button/ButtonAdd/ButtonAdd.scss
@@ -1,0 +1,41 @@
+.Library__Page-Button {
+  width: 60%;
+  height: 50px;
+  border: 1px solid white;
+  margin-top: 22px;
+  display: inline-block;
+  cursor: pointer;
+  color: #000000;
+  text-transform: uppercase;
+  font-size: 12px;
+  letter-spacing: 0.15rem;
+  transition: all 0.5s;
+  position: relative;
+  overflow: hidden;
+  z-index: 1;
+  &:after {
+    content: "";
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    background-color: #fff;
+    z-index: -2;
+  }
+  &:before {
+    content: "";
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    width: 0%;
+    height: 100%;
+    background-color: darken(white, 15%);
+    transition: all 0.3s;
+    z-index: -1;
+  }
+  &:hover {
+    color: #fff;
+    &:before {
+      width: 100%;
+    }
+  }
+}
