+++ client/src/components/UI/Button/ButtonMyBook/MyBookButton.scss
@@ -1,0 +1,32 @@
+.MyBook__List-Button {
+  text-align: center;
+  margin-top: 30px;
+}
+
+.MyBook__List-Button-Status {
+  width: 130px;
+  height: 30px;
+  font-family: "Roboto", sans-serif;
+  font-weight: 300;
+  background-color: #d0d0e5;
+  border: 1px solid #d9d9d9;
+  border-radius: 3px;
+  transition: 0.5s linear;
+  &:hover {
+    background-color: #ababc9;
+  }
+}
+
+.MyBook__List-Button-Status-Disabled {
+  width: 130px;
+  height: 30px;
+  font-family: "Roboto", sans-serif;
+  font-weight: 300;
+  background-color: #8282db;
+  border: 1px solid #d9d9d9;
+  border-radius: 3px;
+  transition: 0.5s linear;
+  &:hover {
+    background-color: #ababc9;
+  }
+}
