+++ client/src/components/UI/Input/InputSearched.scss
@@ -1,0 +1,15 @@
+.Library__Catalog-Input {
+  text-align: center;
+  padding: 20px;
+}
+
+.Library__Catalog-Searched {
+  width: 78%;
+}
+.Library__Catalog-Button {
+  height: 40px;
+}
+
+.Library__Input__Searched {
+  height: 40px;
+}
