+++ client/src/components/UserList/UserList.scss
@@ -1,0 +1,25 @@
+.Library__Page-Users {
+  max-width: 75%;
+  margin: 0 auto;
+  h1 {
+    margin-bottom: 30px;
+  }
+  p {
+    color: #000000;
+    margin-top: 10px;
+  }
+}
+
+.Library__Page-AddUser {
+  margin-top: 30px;
+}
+
+.Library__Page-Static {
+  display: flex;
+  justify-content: center;
+  gap: 80px;
+
+  h3 {
+    font-size: 21px;
+  }
+}
