+++ client/src/pages/AuthPage/AuthPage.scss
@@ -1,0 +1,130 @@
+* {
+  font-family: Helvetica-light;
+}
+
+.main-block {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: 90vh;
+  width: 100%;
+}
+
+.title {
+  margin: 20px 0px 30px 0px;
+  font-family: "Roboto", sans-serif;
+  font-size: 28px;
+  letter-spacing: 3px;
+  color: #110a0a;
+  text-align: center;
+  margin-bottom: 50px;
+}
+
+.form-box {
+  position: relative;
+  width: 400px;
+  height: 450px;
+  background-color: transparent;
+  border: 2px solid #c5c5e4;
+  border-radius: 20px;
+  backdrop-filter: blur(15px);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-family: "Roboto", sans-serif;
+}
+
+.inputbox {
+  position: relative;
+  margin: 30px 0;
+  width: 310px;
+  border-bottom: 2px solid #c5c5e4;
+}
+
+.inputbox label {
+  position: absolute;
+  top: 10px;
+  left: 5px;
+  transform: translateY(-50%);
+  color: #1b1616;
+  font-size: 1em;
+  pointer-events: none;
+  transition: 0.5s;
+}
+
+input:focus ~ label,
+input:focus ~ :valid {
+  top: 0px;
+}
+
+.label__href {
+  margin-left: 10px;
+}
+
+.inputbox input {
+  width: 100%;
+  height: 50px;
+  background: transparent;
+  border: none;
+  outline: none;
+  font-size: 1em;
+  padding: 0 35px 0 5px;
+  color: #110f0f;
+}
+
+.forget {
+  margin: -10px 0 15px;
+  font-size: 0.9em;
+  color: #1a1414;
+  display: flex;
+  justify-content: center;
+  font-family: "Roboto", sans-serif;
+}
+
+.forget label input {
+  margin-right: 3px;
+}
+
+.forget label a {
+  color: #080707;
+  text-decoration: none;
+}
+
+.forget label a:hover {
+  text-decoration: underline;
+}
+
+.btn-log {
+  width: 100%;
+  height: 40px;
+  border-radius: 40px;
+  background: #d0d0e5;
+  border: none;
+  outline: none;
+  cursor: pointer;
+  font-size: 1em;
+  font-weight: 600;
+  transition: 0.5s all;
+  color: #fff;
+}
+.btn-log:hover {
+  background-color: #c5c5e4;
+}
+
+.register {
+  font-size: 0.9em;
+  color: #fff;
+  text-align: center;
+  margin: 25px 0 10px;
+}
+
+.register p a {
+  text-decoration: none;
+  color: #fff;
+  font-weight: 600;
+  margin-left: 10px;
+}
+
+.register p a:hover {
+  text-decoration: underline;
+}
