+++ client/src/pages/LibraryCatalog/LibraryCatalog.scss
@@ -1,0 +1,22 @@
+.Library__Catalog-Container {
+  max-width: 70%;
+  margin: 0 auto;
+  font-family: "Roboto", sans-serif;
+}
+
+.Library__Catalog-Title {
+  font-family: "Roboto", sans-serif;
+  text-align: center;
+  font-weight: 400;
+  font-size: 42px;
+  margin-bottom: 30px;
+  color: #1a1a1a;
+}
+
+.Library__Catalog-Error {
+  margin-top: 350px;
+}
+
+.Library__Catalog-Loading {
+  margin-top: 350px;
+}
