+++ client/src/pages/LibraryPage/LibraryPage.scss
@@ -1,0 +1,62 @@
+.Library__Page-Container {
+  max-width: 75%;
+  margin: 0 auto;
+  font-family: "Roboto", sans-serif;
+  text-align: center;
+  font-weight: 300;
+  h1 {
+    font-size: 35px;
+    font-weight: 400;
+  }
+  hr {
+    margin-bottom: 30px;
+    margin-top: 30px;
+  }
+
+  h2 {
+    font-size: 23px;
+    margin-bottom: 30px;
+  }
+}
+
+.Library__Page-Descr-Bold {
+  margin-top: 30px;
+  margin-bottom: 30px;
+  font-weight: 300;
+}
+
+.Library__Page-Button-Group {
+  margin-bottom: 30px;
+}
+
+.Library__Page-Button-Status {
+  width: 110px;
+  height: 30px;
+  font-family: "Roboto", sans-serif;
+  font-weight: 300;
+  background-color: #d0d0e5;
+  border: 1px solid #d9d9d9;
+  border-radius: 3px;
+  transition: 0.5s linear;
+  &:hover {
+    background-color: #ababc9;
+  }
+}
+
+.Library__Page-Button-Status-Disabled {
+  width: 110px;
+  height: 30px;
+  font-family: "Roboto", sans-serif;
+  font-weight: 300;
+  background-color: #8282db;
+  border: 1px solid #d9d9d9;
+  border-radius: 3px;
+  transition: 0.5s linear;
+  &:hover {
+    background-color: #8282db;
+  }
+}
+
+.Library__Page-Main-Title {
+  margin-bottom: 30px;
+}
