+++ client/src/pages/MyBooksPage/MyBooks.scss
@@ -1,0 +1,34 @@
+.MyBooks__Title {
+  font-family: "Roboto", sans-serif;
+  text-align: center;
+  font-weight: 400;
+  font-size: 42px;
+  margin-bottom: 30px;
+  color: #1a1a1a;
+}
+
+.MyBooks__Container {
+  max-width: 75%;
+  margin: 0 auto;
+  font-family: "Roboto", sans-serif;
+  text-align: start;
+  font-weight: 300;
+  h2 {
+    font-size: 18px;
+    font-weight: 400;
+    text-align: center;
+    margin-bottom: 10px;
+  }
+  hr {
+    margin-bottom: 30px;
+    margin-top: 30px;
+  }
+}
+
+.MyBooks__Button-Static {
+  text-align: center;
+}
+
+.MyBooks__Button {
+  width: 80%;
+}
