+++ client/src/pages/StaticPage/StaticPage.scss
@@ -1,0 +1,23 @@
+.Static__Container {
+  max-width: 75%;
+  margin: 0 auto;
+  font-family: "Roboto", sans-serif;
+  text-align: start;
+  font-weight: 300;
+  h1 {
+    font-family: "Roboto", sans-serif;
+    text-align: center;
+    font-weight: 400;
+    font-size: 42px;
+    margin-bottom: 30px;
+    color: #1a1a1a;
+  }
+  hr {
+    margin-bottom: 30px;
+    margin-top: 30px;
+  }
+}
+
+.Removed__Static {
+  text-align: center;
+}
